Tests: Shared test doubles
