+++ tests/__init__.py
@@ -1,3 +1,4 @@
 """Tests for samplemaker."""
 
 from . import dummy as dummy
+from . import fakes as fakes
